+++ packages/editor/packages/editor-state/src/effects/codeBlocks/blockTypeUpdater.ts
@@ -1,6 +1,6 @@
 import { StateManager } from '@8f4e/state-manager';
 
-import { getBlockType } from '../../helpers/codeParsers';
+import { getBlockType } from '../../helpers/codeParsers/codeParsers';
 
 import type { CodeBlockGraphicData, EventDispatcher, State } from '../../types';
