+++ packages/editor/packages/editor-state/src/effects/codeBlocks/codeBlockCreator.ts
@@ -1,7 +1,7 @@
 import instructionParser from './codeBlockDecorators/instructionParser';
 
 import { EventDispatcher } from '../../types';
-import { getModuleId } from '../../helpers/codeParsers';
+import { getModuleId } from '../../helpers/codeParsers/codeParsers';
 
 import type { CodeBlockGraphicData, State } from '../../types';
