+++ packages/editor/packages/editor-state/src/effects/codeBlocks/codeBlockDecorators/bufferPlotters/updateGraphicData.ts
@@ -1,6 +1,6 @@
 import { parseBufferPlotters } from './codeParser';
 
-import { gapCalculator } from '../../../../helpers/editor';
+import { gapCalculator } from '../../../../helpers/codeEditing/editor';
 import resolveMemoryIdentifier from '../../../../helpers/resolveMemoryIdentifier';
 
 import type { CodeBlockGraphicData, State } from '../../../../types';
