+++ packages/editor/packages/editor-state/src/effects/codeBlocks/codeBlockDecorators/errorMessages/errorMessages.ts
@@ -1,4 +1,4 @@
-import { gapCalculator } from '../../../../helpers/editor';
+import { gapCalculator } from '../../../../helpers/codeEditing/editor';
 
 import type { CodeBlockGraphicData, State } from '../../../../types';
