+++ packages/editor/packages/editor-state/src/effects/codeBlocks/codeBlockDecorators/outputs/updateGraphicData.ts
@@ -1,7 +1,7 @@
 import { parseOutputs } from './codeParser';
 
-import { gapCalculator } from '../../../../helpers/editor';
-import { getModuleId } from '../../../../helpers/codeParsers';
+import { gapCalculator } from '../../../../helpers/codeEditing/editor';
+import { getModuleId } from '../../../../helpers/codeParsers/codeParsers';
 
 import type { CodeBlockGraphicData, Output, State } from '../../../../types';
