+++ packages/editor/packages/editor-state/src/effects/codeBlocks/codeBlockDecorators/pianoKeyboard/parsePressedKeys.ts
@@ -1,4 +1,4 @@
-import { parseCode } from '../../../../helpers/multiLineCodeParser';
+import { parseCode } from '../../../../helpers/codeParsers/multiLineCodeParser';
 
 export function parsePressedKeys(code: string[], pressedKeysListMemoryId: string, startingNumber: number) {
 	const pressedKeys = new Set<number>();
