+++ packages/editor/packages/editor-state/src/effects/codeBlocks/codeBlockDecorators/switches/updateGraphicData.ts
@@ -1,6 +1,6 @@
 import { parseSwitches } from './codeParser';
 
-import { gapCalculator } from '../../../../helpers/editor';
+import { gapCalculator } from '../../../../helpers/codeEditing/editor';
 
 import type { CodeBlockGraphicData, State } from '../../../../types';
