+++ packages/editor/packages/editor-state/src/testing.ts
@@ -18,4 +18,4 @@ export {
 	createMockViewport,
 	createMockEventDispatcher,
 	createMockState,
-} from './helpers/testUtils';
+} from './helpers/testingUtils/testUtils';
